requirements.txt
pytest==4.5.0
PyMySQL==0.9.3
allure-pytest==2.8.6
requests==2.20.0
PyYAML==5.3.1
redis==4.1.0
rocketmq-client-python==2.0.0rc1
elasticsearch==7.10.0